Update Central Script Logic
This is the central control for the different modes of rope lighting. It begins in 'standby', checking for frequency data outside the passive baseline from the WebSocket server. If it moves outside the baseline, the code passes the pattern to 'boot'. Once boot has cycled for 5 seconds, it auto-passes to 'interactive-passive'. This mode will watch for frequency data above or below the baseline, and pass to 'interactive-forwards' and 'interactive-backwards', respectively. After 60 uninterrupted seconds in interactive-passive, the central code passes to 'shutdown', which lasts for 5 seconds and then auto-passes back to 'standby'.

Notes:
- passive frequency set to 100hz, likely needs adjusting to fit a range
- WebSocket server data is a placeholder and needs updating with actual connection info
- naming conventions for patterns need to match up with the pattern library in Pixelblaze.
- given that we have other lights involved in these sequences, this central logic will need to be blended in to fit several more patterns